requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
xgboost==2.0.3
shap==0.45.1
joblib==1.4.2
pymongo==4.7.2
gunicorn==22.0.0
